Refactored tests to run without internet or fred api key
Introduced a variable fake_fred_call. If fake_fred_call is False, it
behaves the way it used to and makes an actual call to Fred; on the
other hand, if fake_fred_call is True (default value), the response
from Fred is faked by unittest.mock so that tests behave consistently.

Also:
 - Split tests among more methods.
 - Made test_search easier to read and changed the order_by clause from
 popularity to series_id so that the result do not change when running
 the tests against FRED.
 - Added .travis.yml and requirements.txt.
 - Listed the test_suite in setup.py so one can run python setup.py test to run the tests.
 - Added Travis badge to README
 - moved root_url from the object to the class level.
 - cleaned up unnecessary & at the end of the urls.
 - factored pandas import at the begining of the file.