Fixed test_search
Could reproduce the problem on Ubuntu only with Python 3.4. It just
seemed to crash. Removed the use of zip function seem to do the trick.

The test looks better anyway with a single assert and no loop. The only
cost is what looks like unnecessary spaces at the end of the line.